Add options for configuration file location.
This commit includes several changes: the configuration does not know
itself anymore where it was loaded from, the ConfigurationLocator is now
used for that. The preferences page does not need access to the
configuration anymore. Finally, the lock file has been removed; now, when
saving a configuration file would overwrite an existing configuration file,
the user is asked for permission.